Fix SongCollection remove, copyArray methods
Prior to this change, the remove method would enter an infinite loop and
copyArray wasn't functional enough for the remove method. Java has
in-built System.arraycopy method which is preferred.

This change replaces copyArray with System.arraycopy and fixes the logic
inside the remove method. Furthermore remove is provided with a check to
prevent crashing.